kimpt142 commit fix delete request violation 21/07

+++ project/src/main/java/com/example/webDemo3/service/impl/manageEmulationServiceImpl/ViolationOfClassServiceImpl.java
@@ -250,12 +250,22 @@ public class ViolationOfClassServiceImpl implements ViolationOfClassService {
             message = Constant.REQUEST_ID_NULL;
         }
 
+        ViolationClassRequest violationClassRequest = violationClassRequestRepository.findById(requestId).orElse(null);
+        if(violationClassRequest == null){
+            message = Constant.VIOLATIONREQUEST_DELETED;
+            return message;
+        }
+
+        if(violationClassRequest.getStatusChange() == 1 || violationClassRequest.getStatusChange() == 2){
+            message = Constant.REJECT_DELETE_REQUEST_FAIL;
+            return message;
+        }
+
         try {
             violationClassRequestRepository.deleteById(requestId);
         }
         catch (Exception e){
-            message.setMessageCode(1);
-            message.setMessage(e.toString());
+            message = Constant.DELETE_REQUEST_FAIL;
             return message;
         }
